src/webapp/requirements.txt
loguru==0.7.3
fastapi
python-multipart==0.0.19
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.10.6
argon2-cffi==23.1.0
pipecat-ai==0.0.69
beanie==1.29.0
motor==3.7.1
jinja2==3.1.4
pycryptodome
cryptography==44.0.0